PyTorch Framework/requirements.txt
tqdm
pytorch
numpy
pandas
prettytable